# Feature: 100x performance boost to Rust HTTP server