--- a/apps/rusti/Cargo.toml
+++ b/apps/rusti/Cargo.toml
@@ -1,2 +1,4 @@
 [workspace]
-members = ["cli","server", "core"]
+members = ["cli", "server", "core"]
+
+resolver = "2"