--- a/apps/rusti/core/src/lib.rs
+++ b/apps/rusti/core/src/lib.rs
@@ -1,60 +1,51 @@
-use rand::{
-    distributions::{Distribution, Uniform},
-    rngs::ThreadRng,
-    Rng,
-};
+use rand::Rng;
+use rand::RngCore;
 use std::{collections::HashMap, io};
 
 pub fn combine_zip(words: &[String], separators: &[String]) -> String {
-    let mut i = 0;
-    let separators_len = separators.len();
     let mut result = String::with_capacity(
         // Pre-allocate since the size is known ahead of time
         words.iter().map(|x| x.len()).sum::<usize>()
             + separators.iter().map(|x| x.len()).sum::<usize>(),
     );
 
-    for word in words {
+    for (i, word) in words.iter().enumerate() {
         result.push_str(word);
-        if i < separators_len {
-            result.push_str(&separators[i]);
+        if let Some(sep) = separators.get(i) {
+            result.push_str(sep);
         }
-        i += 1;
     }
     result
 }
 
 pub fn make_words(count: usize, word_list: &HashMap<String, String>) -> Vec<String> {
     make_wl_keys(count)
-        .iter()
-        .flat_map(|k| match word_list.get(k) {
-            Some(word) => Some(word.clone()),
-            _ => None,
-        })
+        .into_iter()
+        .filter_map(|k| word_list.get(&k).cloned())
         .collect()
-}
-
-struct WordListKeys {
-    rng: ThreadRng,
-}
-
-impl Iterator for WordListKeys {
-    type Item = Vec<String>;
-
-    fn next(&mut self) -> Option<Self::Item> {
-        let die = Uniform::new_inclusive(1, 6);
-        let roll = die.sample_iter(&mut self.rng);
-        let keys: Vec<String> = roll.take(5).map(|x| x.to_string()).collect();
-        Some(keys)
-    }
 }
 
 fn make_wl_keys(count: usize) -> Vec<String> {
-    let rng = rand::thread_rng();
-    WordListKeys { rng }
-        .take(count)
-        .map(|roll_result| roll_result.join(""))
-        .collect()
+    const KEY_LENGTH: usize = 5;
+    const DIGIT_RANGE: u8 = 6;
+
+    let total_byte_count = count * KEY_LENGTH;
+    let mut rng = rand::thread_rng();
+    let mut bytes = vec![0u8; total_byte_count];
+    rng.fill_bytes(&mut bytes);
+
+    let mut keys: Vec<String> = Vec::with_capacity(count);
+    let mut key = String::with_capacity(KEY_LENGTH);
+
+    for chunk in bytes.chunks_exact(KEY_LENGTH) {
+        for &byte in chunk {
+            let digit = ((byte % DIGIT_RANGE) + 1) as u32;
+            key.push(std::char::from_digit(digit, 10).unwrap());
+        }
+        keys.push(key.clone());
+        key.clear();
+    }
+    keys
 }
 
 const SEPARATOR_OPTS: [&'static str; 12] =
@@ -62,26 +53,15 @@ const SEPARATOR_OPTS: [&'static str; 12] =
 const SEPARATOR_OPTS_LEN: usize = SEPARATOR_OPTS.len();
 
 pub fn make_separators(count: usize, separator_kind: &str) -> Vec<String> {
-    let sep_count = count - 1;
+    let sep_count = count.saturating_sub(1);
     match separator_kind {
         "random" => {
             let mut rng = rand::thread_rng();
             (0..sep_count)
-                .into_iter()
-                .filter_map(|_| {
-                    let random_index = rng.gen_range(0..SEPARATOR_OPTS_LEN);
-                    match SEPARATOR_OPTS.get(random_index) {
-                        Some(sep) => Some(sep.clone().to_owned()),
-                        _ => None,
-                    }
-                })
+                .map(|_| SEPARATOR_OPTS[rng.gen_range(0..SEPARATOR_OPTS_LEN)].to_owned())
                 .collect()
         }
-        ch => ch
-            .repeat(sep_count)
-            .split("")
-            .filter_map(|x| if x != "" { Some(x.to_owned()) } else { None })
-            .collect(),
+        ch => vec![ch.to_owned(); sep_count],
     }
 }