Never got around to benchmarking the rust HTTP server and it turns out I wrote some very inefficient rust 😆 

Original oha benchmark result:
```shell
oha http://localhost:8080/gen\?count\=11\&sep\=.\&fmt\=text -z 10s
Summary:
  Success rate:	100.00%
  Total:	10.0016 secs
  Slowest:	0.1662 secs
  Fastest:	0.0142 secs
  Average:	0.0437 secs
  Requests/sec:	1138.6219

  Total data:	972.65 KiB
  Size/request:	87 B
  Size/sec:	97.25 KiB

Response time histogram:
  0.014 [1]    |
  0.029 [4152] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.045 [2161] |■■■■■■■■■■■■■■■■
  0.060 [2081] |■■■■■■■■■■■■■■■■
  0.075 [1651] |■■■■■■■■■■■■
  0.090 [942]  |■■■■■■■
  0.105 [318]  |■■
  0.121 [65]   |
  0.136 [13]   |
  0.151 [2]    |
  0.166 [2]    |

Response time distribution:
  10.00% in 0.0167 secs
  25.00% in 0.0207 secs
  50.00% in 0.0404 secs
  75.00% in 0.0611 secs
  90.00% in 0.0778 secs
  95.00% in 0.0863 secs
  99.00% in 0.1019 secs
  99.90% in 0.1262 secs
  99.99% in 0.1523 secs


Details (average, fastest, slowest):
  DNS+dialup:	0.0016 secs, 0.0013 secs, 0.0026 secs
  DNS-lookup:	0.0001 secs, 0.0000 secs, 0.0006 secs

Status code distribution:
  [200] 11388 responses
```

And now after all the changes:
```shell
 oha http://localhost:8080/gen\?c\=11\&s\=.\&fmt\=txt -z 10s
Summary:
  Success rate:	100.00%
  Total:	10.0018 secs
  Slowest:	0.0098 secs
  Fastest:	0.0000 secs
  Average:	0.0004 secs
  Requests/sec:	126693.5683

  Total data:	105.00 MiB
  Size/request:	86 B
  Size/sec:	10.50 MiB

Response time histogram:
  0.000 [1]       |
  0.001 [1195695] |■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■■
  0.002 [58294]   |■
  0.003 [8845]    |
  0.004 [2544]    |
  0.005 [1076]    |
  0.006 [396]     |
  0.007 [218]     |
  0.008 [55]      |
  0.009 [27]      |
  0.010 [11]      |

Response time distribution:
  10.00% in 0.0001 secs
  25.00% in 0.0002 secs
  50.00% in 0.0003 secs
  75.00% in 0.0004 secs
  90.00% in 0.0008 secs
  95.00% in 0.0011 secs
  99.00% in 0.0020 secs
  99.90% in 0.0042 secs
  99.99% in 0.0066 secs


Details (average, fastest, slowest):
  DNS+dialup:	0.0030 secs, 0.0017 secs, 0.0047 secs
  DNS-lookup:	0.0002 secs, 0.0000 secs, 0.0011 secs

Status code distribution:
  [200] 1267162 responses
```

Bunch of little details went into this:
- Read the large wordlist 1 time at startup for each workers instead of on each request. This made the biggest difference. The JSON file is ~140kb and reading it into a hash map each request was super expensive.
- Refactor lib.rs 
  - Access randomness source 1 time per key list generation
  - Use simpler/terser methods for iteration
  - Rewrite `make_wl_keys` to match similar implementation to source ReScript. Pre-allocates an array of random bytes instead of the custom iterator.


Other changes:
- Update rust deps
- Set new resolver in cargo
- Align on HTTP api params format with fallback when separator is not provided